+++ admin-mbg/src/main/java/com/myproject/admin/model/CmsCourseItemExample.java
@@ -565,52 +565,52 @@ public class CmsCourseItemExample {
             return (Criteria) this;
         }
 
-        public Criteria andTeacherIdEqualTo(Integer value) {
+        public Criteria andTeacherIdEqualTo(Long value) {
             addCriterion("teacher_id =", value, "teacherId");
             return (Criteria) this;
         }
 
-        public Criteria andTeacherIdNotEqualTo(Integer value) {
+        public Criteria andTeacherIdNotEqualTo(Long value) {
             addCriterion("teacher_id <>", value, "teacherId");
             return (Criteria) this;
         }
 
-        public Criteria andTeacherIdGreaterThan(Integer value) {
+        public Criteria andTeacherIdGreaterThan(Long value) {
             addCriterion("teacher_id >", value, "teacherId");
             return (Criteria) this;
         }
 
-        public Criteria andTeacherIdGreaterThanOrEqualTo(Integer value) {
+        public Criteria andTeacherIdGreaterThanOrEqualTo(Long value) {
             addCriterion("teacher_id >=", value, "teacherId");
             return (Criteria) this;
         }
 
-        public Criteria andTeacherIdLessThan(Integer value) {
+        public Criteria andTeacherIdLessThan(Long value) {
             addCriterion("teacher_id <", value, "teacherId");
             return (Criteria) this;
         }
 
-        public Criteria andTeacherIdLessThanOrEqualTo(Integer value) {
+        public Criteria andTeacherIdLessThanOrEqualTo(Long value) {
             addCriterion("teacher_id <=", value, "teacherId");
             return (Criteria) this;
         }
 
-        public Criteria andTeacherIdIn(List<Integer> values) {
+        public Criteria andTeacherIdIn(List<Long> values) {
             addCriterion("teacher_id in", values, "teacherId");
             return (Criteria) this;
         }
 
-        public Criteria andTeacherIdNotIn(List<Integer> values) {
+        public Criteria andTeacherIdNotIn(List<Long> values) {
             addCriterion("teacher_id not in", values, "teacherId");
             return (Criteria) this;
         }
 
-        public Criteria andTeacherIdBetween(Integer value1, Integer value2) {
+        public Criteria andTeacherIdBetween(Long value1, Long value2) {
             addCriterion("teacher_id between", value1, value2, "teacherId");
             return (Criteria) this;
         }
 
-        public Criteria andTeacherIdNotBetween(Integer value1, Integer value2) {
+        public Criteria andTeacherIdNotBetween(Long value1, Long value2) {
             addCriterion("teacher_id not between", value1, value2, "teacherId");
             return (Criteria) this;
         }
